Bug fixes and updates
Corrected issues with newlines and other unwanted characvters in user-input fields, whichcan mess with the CSV format.

Also updated the script to drop patron names from API data, and to only store HASHES of email addresses instead of plaintext ones. This is intended to help preserve patron privacy, since we do not need to know this information.